Renewal in renew scheduler
Instead of using a durable function to coordinate certificate renewal
across many functions, all of the certificates are renewed in the
weekly schedule function instead. This simplifies the overall structure
of the function app, and reduces the amount of traffic internally in
Azure. Specifically, the function app polls less Azure queues.